feat(routes)!: reorganize some routes (oauth vs apptoken)
