+++ config/routes.yaml
@@ -22,8 +22,12 @@ logout:
   # we don't need to specify a controller, since `security.yaml` identifies this route-key as logout, and does magic
 
 account:
-  resource: 'routes/accounts.yaml'
+  resource: 'routes/user/account.yaml'
   prefix: /account
+
+appToken:
+  resource: 'routes/user/appToken.yaml'
+  prefix: /appToken
 
 api:
   utf8: true
@@ -32,5 +36,5 @@ api:
 
 oauth:
   utf8: true
-  resource: 'routes/api/routes.yaml'
+  resource: 'routes/user/oauth.yaml'
   prefix: /㊙
