+++ src/Controller/User/OAuthController.php
@@ -2,122 +2,22 @@
 
 namespace App\Controller\User;
 
-use App\Dto\AppTokenDto;
-use App\Entity\AppToken;
 use App\Entity\User;
 use App\Exception\OAuthException;
 use App\Exception\UserModificationException;
 use App\Security\TokenAuthority;
-use App\Utility\QoL;
-use App\Utility\UserInputStrings;
-use DateTime;
 use Exception;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\Security\Core\Exception\InvalidCsrfTokenException;
-use Symfony\Component\Security\Csrf\CsrfToken;
-use Symfony\Component\Security\Csrf\CsrfTokenManagerInterface;
 
 /**
  * @method User|null getUser()
  */
 class OAuthController extends AbstractController
 {
-    private const CRSF_GENERATE_APP_TKN = 'generate-token';
-
-    private CsrfTokenManagerInterface $csrfTokenManager;
-
-    public function __construct(
-        CsrfTokenManagerInterface $csrfTokenManager
-    ) {
-        $this->csrfTokenManager = $csrfTokenManager;
-    }
-
-    /**
-     * @SuppressWarnings(PHPMD.StaticAccess)
-     */
-    public function appTokenList(): Response
-    {
-        $user = $this->getUser();
-        if (!$user instanceof User) {
-            throw new Exception('User is not logged in');
-        }
-
-        $appTokens = array_reduce(
-            $user->getAppTokens()->toArray(),
-            fn(array $tokens, AppToken $token) => QoL::arrPush($tokens, new AppTokenDto($token)),
-            /** @var AppTokenDto[] */
-            []
-        );
-
-        return $this->render('account/app-tokens.html.twig', [
-            'tokens' => $appTokens,
-            'crsfId' => self::CRSF_GENERATE_APP_TKN,
-        ]);
-    }
-
-    /**
-     * @SuppressWarnings(PHPMD.StaticAccess)
-     */
-    public function createAppToken(Request $request, TokenAuthority $tokenAuthority): Response
-    {
-        $crsfToken = $request->request->get('_csrf_token');
-
-        if (!$this->csrfTokenManager->isTokenValid(new CsrfToken(self::CRSF_GENERATE_APP_TKN, $crsfToken))) {
-            throw new InvalidCsrfTokenException();
-        }
-
-        // We want to remove the token from the session, so the user can't just refresh the page to create new tokens.
-        $this->csrfTokenManager->removeToken(self::CRSF_GENERATE_APP_TKN);
-
-        $tokenName = UserInputStrings::trimMb4String($request->request->get('name') ?? '');
-        $tokenExpiration = $request->request->get('expiration');
-        $tokenExpiration = empty($tokenExpiration) ? null : new DateTime($tokenExpiration);
-
-        if (mb_strlen($tokenName) === 0) {
-            return $this->redirectToRoute('oauth.app.list', [], Response::HTTP_TEMPORARY_REDIRECT);
-        }
-
-        $user = $this->getUser();
-        if (!$user instanceof User) {
-            throw new Exception('User is not logged in');
-        }
-
-        $tokenEntity = $tokenAuthority->createToken($user, $tokenName, $tokenExpiration);
-
-        return $this->render('account/new-app-token.html.twig', [
-            'name' => $tokenEntity->getName(),
-            'token' => $tokenEntity->getAuthorizationToken(),
-        ]);
-    }
-
-    public function deleteAppToken(string $uuid): JsonResponse
-    {
-        $user = $this->getUser();
-        if (!$user instanceof User) {
-            throw new Exception('User is not logged in');
-        }
-
-        /** @var AppToken|false */
-        /** @psalm-suppress UnnecessaryVarAnnotation */
-        $token = $user->getAppTokens()->filter(fn(AppToken $token) => $token->getUuid() === $uuid)->first();
-
-        if ($token instanceof AppToken) {
-            $user->removeAppToken($token);
-            $entityManager = $this->getDoctrine()->getManager();
-            $entityManager->persist($user);
-            $entityManager->flush();
-            $entityManager->clear();
-
-            return new JsonResponse();
-        }
-
-        return new JsonResponse([], Response::HTTP_NOT_FOUND);
-    }
-
     public function generateOAuth(): Response
     {
         $authenticatedUser = $this->getUser();
@@ -173,6 +73,12 @@ class OAuthController extends AbstractController
 
         // From here, need to provide a refresh token, and an access token.
         // Need to build new entities.
+        http_build_query([
+            'access_token' => 'AQXNnd2kXITHELmWblJigbHEuoFdfRhOwGA0QNnumBI8XOVSs0HtOHEU-wvaKrkMLfxxaB1O4poRg2svCWWgwhebQhqrETYlLikJJMgRAvH1ostjXd3DP3BtwzCGeTQ7K9vvAqfQK5iG_eyS-q-y8WNt2SnZKZumGaeUw_zKqtgCQavfEVCddKHcHLaLPGVUvjCH_KW0DJIdUMXd90kWqwuw3UKH27ki5raFDPuMyQXLYxkqq4mYU-IUuZRwq1pcrYp1Vv-ltbA_svUxGt_xeWeSxKkmgivY_DlT3jQylL44q36ybGBSbaFn-UU7zzio4EmOzdmm2tlGwG7dDeivdPDsGbj5ig',
+            'expires_in' => 86400,
+            'refresh_token' => 'AQWAft_WjYZKwuWXLC5hQlghgTam-tuT8CvFej9-XxGyqeER_7jTr8HmjiGjqil13i7gMFjyDxh1g7C_G1gyTZmfcD0Bo2oEHofNAkr_76mSk84sppsGbygwW-5oLsb_OH_EXADPIFo0kppznrK55VMIBv_d7SINunt-7DtXCRAv0YnET5KroQOlmAhc1_HwW68EZniFw1YnB2dgDSxCkXnrfHYq7h63w0hjFXmgrdxeeAuOHBHnFFYHOWWjI8sLLenPy_EBrgYIitXsAkLUGvZXlCjAWl-W459feNjHZ0SIsyTVwzAQtl5lmw1ht08z5Du-RiQahQE0sv89eimHVg9VSNOaTvw',
+            'refresh_token_expires_in' =>  525600,
+        ]);
 
         return new RedirectResponse($returnUrl);
     }
