+++ templates/account/new-app-token.html.twig
@@ -25,7 +25,7 @@
         </p>
     </div>
     <div>
-        <a href="{{ path('oauth.app.list') }}">Return to app token list</a>
+        <a href="{{ path('appToken.list') }}">Return to app token list</a>
     </div>
     <p id="outbox"></p>
 {% endblock %}
