+++ templates/base.html.twig
@@ -29,7 +29,7 @@
                             <a href="{{ path('account.edit') }}">Profile</a>
                         </li>
                         <li tabindex="0">
-                            <a href="{{ path('oauth.app.list') }}">App Tokens</a>
+                            <a href="{{ path('appToken.list') }}">App Tokens</a>
                         </li>
                         <li tabindex="0">
                             <a href="{{ path('logout') }}">Logout</a>
